requirements.txt
pytest==6.0.2
pytest-html==3.0.0
pytest-xdist==2.1.0
pytest-rerunfailures==9.1.1
allure-pytest==2.8.18
requests==2.24.0
requests-toolbelt==0.9.1
simplejson==3.17.2
ruamel.yaml==0.16.12
PyYAML==5.3.1
configparser==5.0.0
python-dateutil==2.8.0